Fix BandStars being written as Int32

+++ YARG.Core/Replays/ReplaySerializer.Sections.cs
@@ -59,7 +59,7 @@ namespace YARG.Core.Replays
                 writer.Write(metadata.ArtistName);
                 writer.Write(metadata.CharterName);
                 writer.Write(metadata.BandScore);
-                writer.Write((int) metadata.BandStars);
+                writer.Write((byte) metadata.BandStars);
                 writer.Write(metadata.ReplayLength);
                 writer.Write(metadata.Date.ToBinary());
                 metadata.SongChecksum.Serialize(writer);
